+++ compiler/src/main/java/info/teksol/mc/util/CollectionUtils.java
@@ -162,4 +162,12 @@ public class CollectionUtils {
             return List.copyOf(list);
         }
     }
+
+    public static <T> boolean identicalInstances(List<T> list1, List<T> list2) {
+        if (list1.size() != list2.size()) return false;
+        for (int i = 0; i < list1.size(); i++) {
+            if (list1.get(i) != list2.get(i)) return false;
+        }
+        return true;
+    }
 }